Completed Phase 4
Earning $50 means you win, which sends you to a WinState that has a
retry or main menu button (the button images used are place holders) the
same applies for the lose state. The amount of people served and left
are also displayed on both states as well as the GameState.