Phase 2 Completed
Added feature to ringCustomer() to give all clients a fourth of their
patience back(with OGpatience being the maximum bound), if matched is
true.
Altered the ringCustomer() method to accept client as a parameter.
Implemented if statements to accept numbers as input for PlateCounter
and serve the customer whose number is pressed.
Added RadioCounter class and implemented it into the game by increasing
width of the game by one counter size. Added variables to hold values
for a random number of ticks and time window for counter effect. The
counter will set all client's patience to their maximum when interacted
with during the time window.